remove mined ore move to map module
