--- src/app/component/help/help.component.html
@@ -1,1 +1,0 @@
-<p>help works!</p>
