--- src/app/component/help/help.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'dumap-help',
-  templateUrl: './help.component.html',
-  styleUrls: ['./help.component.scss']
-})
-export class HelpComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
