--- src/app/component/subtract-mined-ore-dialog/subtract-mined-ore-dialog.component.scss
@@ -1,69 +1,0 @@
-.dialog {
-    position: absolute;
-    width:50%;
-    z-index: 100;
-    background-color: rgb(27,40,48);
-    padding: 5px;
-}
-.centered {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    /* bring your own prefixes */
-    transform: translate(-50%, -50%);
-}
-
-table th, table td{
-    padding: 10px; /* Apply cell padding */
-}
-
-table{    
-    border-spacing: 0px; /* Apply cell spacing */
-    background-color: rgb(13,24,28);
-}
-
-.tileHeader {
-    background-color: rgb(20, 32, 39); 
-    color: white;
-    font-weight: bold;
-}
-
-.leftColumn {
-    background-color: rgb(13,24,28);
-     color: rgb(76,112,124);
-}
-.rightColumn {
-    background-color: rgb(13,24,28);
-    color: rgb(182,222,236);
-}
-
-.map_modal {    
-    position: fixed; /* Stay in place */
-    z-index: 10; /* Sit on top */
-    left: 0;
-    top: 0;
-    width: 100%; /* Full width */
-    height: 100%; /* Full height */
-    overflow: auto; /* Enable scroll if needed */
-    background-color: rgb(0,0,0); /* Fallback color */
-    background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
-}
-
-.isModal {
-      display: none;
-}
-
-.tab {
-    background-color: rgb(13,24,28);
-    border-radius: 10px 10px 0px 0px;
-    border-top: 3px solid rgb(27,40,48);
-    border-left: 3px solid rgb(27,40,48);
-    border-right: 3px solid rgb(27,40,48);
-    color: white;
-    text-align: center;
-    padding: 5px
-}
-
-.tab-active {
-    border-bottom: 3px solid rgb(13,24,28);
-}
