--- src/app/component/subtract-mined-ore-dialog/subtract-mined-ore-dialog.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SubtractMinedOreDialogComponent } from './subtract-mined-ore-dialog.component';
-
-describe('SubtractMinedOreDialogComponent', () => {
-  let component: SubtractMinedOreDialogComponent;
-  let fixture: ComponentFixture<SubtractMinedOreDialogComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ SubtractMinedOreDialogComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SubtractMinedOreDialogComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
