+++ src/app/map/component/add-scan-dialog/add-scan-dialog.component.spec.ts
