+++ src/app/map/component/detail-window/detail-window.component.html
@@ -33,14 +33,14 @@
         </tr>
 
         <tr *ngIf="scan">
-            <td class="rightColumn">&sum;</td><td class="rightColumn" style="text-align: right; color: red">{{difOre(scan) > 0 ? (-difOre(scan) | number:'1.0-0':getUsersLocale('en-US')) : ''}}</td><td class="rightColumn" style="text-align: right;">{{sumOre(scan) | number:'1.0-0':getUsersLocale('en-US')}} L/h</td>
+            <td class="rightColumn">&sum;</td><td class="rightColumn" style="text-align: right;">{{sumOre(scan) | number:'1.0-0':getUsersLocale('en-US')}} L/h</td>
         </tr>
         <tr *ngIf="scan">
-            <td class="rightColumn">Quanta</td><td class="rightColumn" style="text-align: right; color: red">{{difQuanta(scan) > 0 ? (-difQuanta(scan) | number:'1.0-0':getUsersLocale('en-US')) : ''}}</td><td class="rightColumn" style="text-align: right;">~{{sumQuanta(scan) | number:'1.0-0':getUsersLocale('en-US')}} &#295;/h<br>~{{sumQuanta(scan)*7*24 | number:'1.0-0':getUsersLocale('en-US')}} &#295;/wk</td>
+            <td class="rightColumn">Quanta</td><td class="rightColumn" style="text-align: right;">~{{sumQuanta(scan) | number:'1.0-0':getUsersLocale('en-US')}} &#295;/h<br>~{{sumQuanta(scan)*7*24 | number:'1.0-0':getUsersLocale('en-US')}} &#295;/wk</td>
         </tr>
         <ng-container *ngFor="let ore of oreNames">
         <tr  *ngIf="scan?.ores[ore.name]">
-            <td class="rightColumn">{{ore.name}}</td><td class="rightColumn" style="text-align: right; color: red" [title]="minedOreDetail(scan, ore.name)">{{sumOreMined(scan, ore.name) > 0 ? (-sumOreMined(scan, ore.name) | number:'1.0-0':getUsersLocale('en-US')) : ''}}</td><td class="rightColumn" style="text-align: right;">{{scan.ores[ore.name] | number:'1.0-0':getUsersLocale('en-US')}} {{postFix}}</td>
+            <td class="rightColumn">{{ore.name}}</td><td class="rightColumn" style="text-align: right;">{{scan.ores[ore.name] | number:'1.0-0':getUsersLocale('en-US')}} {{postFix}}</td>
         </tr>
         </ng-container>
     </ng-container>
