+++ src/app/map/component/detail-window/detail-window.component.spec.ts
