+++ src/app/map/component/detail-window/detail-window.component.ts
@@ -1,9 +1,9 @@
 import { DatePipe } from '@angular/common';
 import { Component, Inject, OnInit } from '@angular/core';
 import { faCopy } from '@fortawesome/free-solid-svg-icons';
-import { Face } from 'src/app/model/Face';
-import { Scan } from 'src/app/model/Scan';
-import { EventService } from 'src/app/service/event.service';
+import { Face } from 'src/app/map/model/Face';
+import { Scan } from 'src/app/map/model/Scan';
+import { EventService } from 'src/app/map/service/event.service';
 
 @Component({
   selector: 'dumap-detail-window',
@@ -88,60 +88,36 @@ export class DetailWindowComponent implements OnInit {
     return lang;
   }
 
-  minedOreDetail(scan: Scan, oreName: string): string {
-    let out = '';
-    if (scan.minedOre) {
-      const date = new DatePipe(this.getUsersLocale('en-US'));
-      for (const minedOre of scan.minedOre) {
-        if (minedOre.ores[oreName]) {
-          // tslint:disable-next-line:max-line-length
-          out += `${'         '.slice(minedOre.ores[oreName].toLocaleString().length)}${minedOre.ores[oreName]} L ${minedOre.owner} ${date.transform(minedOre.time, 'MMM dd, y HH:mm')}\n`;
-        }
-      }
-    }
-    return out.trim();
+  sumOre(scan: Scan): number  {
+    return Scan.sumOre(scan);
   }
 
-  sumOre(scan: Scan, subtractMined = false): number  {
-    return Scan.sumOre(scan, subtractMined);
-  }
   difOre(scan: Scan): number {
-    return this.sumOre(scan, false) - this.sumOre(scan, true);
+    return this.sumOre(scan) - this.sumOre(scan);
   }
 
-  sumHc(scan: Scan, subtractMined = false): number {
-    return Scan.sumHc(scan, this.oreNames, subtractMined);
+  sumHc(scan: Scan): number {
+    return Scan.sumHc(scan, this.oreNames);
   }
+
   difHC(scan: Scan): number {
-    return this.sumHc(scan, false) - this.sumHc(scan, true);
+    return this.sumHc(scan) - this.sumHc(scan);
   }
 
-  sumQuanta(scan: Scan, subtractMined = false): number {
-    return Scan.sumQuanta(scan, this.oreNames, subtractMined);
+  sumQuanta(scan: Scan): number {
+    return Scan.sumQuanta(scan, this.oreNames);
   }
+
   difQuanta(scan: Scan): number {
-    return this.sumQuanta(scan, false) - this.sumQuanta(scan, true);
-  }
-
-  sumOreMined(scan: Scan, name: string): number {
-    return Scan.sumOreMined(scan, name);
+    return this.sumQuanta(scan) - this.sumQuanta(scan);
   }
 
   onScanClick() {
     // tslint:disable-next-line:max-line-length
     let out = `Planet: ${this.scan.planet}\nTile: ${this.scan.tileId}\nScaned by: ${this.scan.owner}\n${this.posLink()}\n${new DatePipe(this.getUsersLocale('en-US')).transform(this.scan.time, 'MMM dd, y HH:mm')}\n`;
-    const hasMined = this.scan.minedOre && this.scan.minedOre.length > 0;
-    for (let ore of this.oreNames) {
+    for (const ore of this.oreNames) {
       if (this.scan.ores[ore.name]) {
-        const mined = this.sumOreMined(this.scan, ore.name);
         out += `\n${ore.name}:${'            '.slice(ore.name.length)} `;
-        if (hasMined) {
-          if (mined > 0) {
-            out += `${'         '.slice(mined.toLocaleString().length)}-${mined.toLocaleString()} `;
-          } else  {
-            out += '           ';
-          }
-        }
         out += `${'         '.slice(this.scan.ores[ore.name].toLocaleString().length)}${this.scan.ores[ore.name].toLocaleString()}`;
       }
     }
