+++ src/app/map/component/map-component/map-component.component.scss
