+++ src/app/map/component/map-settings/map-settings.component.scss
