+++ src/app/map/component/map-settings/map-settings.component.spec.ts
