+++ src/app/map/component/map/map.component.scss
@@ -1,0 +1,37 @@
+body, div {
+    background-color: rgb(27,40,48);
+    padding: 0 ;
+}
+
+.is-270px {
+    min-width: 270px;
+}
+
+.btn {
+    height: 50px;
+    line-height: 33px;  
+    width: 50px;  
+    font-size: 2em;
+    font-weight: bold;
+    border-radius: 50%;
+    background-color: rgb(20, 32, 39);
+    color: white;
+    text-align: center;
+    cursor: pointer;
+    border: 5px solid rgb(27,40,48);
+  }
+
+.Nav {
+    background-color: rgb(27,40,48); 
+    padding: 0; 
+    min-height: 100%; 
+    height: auto;  
+    overflow: hidden
+}
+.maximizedNav {
+    width: 239px; 
+    background-image: url('/assets/maps.PNG'); 
+}
+.minimizedNav {
+    width: 89px;    
+}
