+++ src/app/map/component/scan-list/scan-list.component.html
