+++ src/app/map/directive/customMax.directive.ts
