+++ src/app/map/directive/customMin.directive.ts
