+++ src/app/map/model/SelectedTile.ts
