+++ src/app/map/model/Settings.ts
