+++ src/app/map/service/event.service.ts
@@ -1,6 +1,5 @@
 import { EventEmitter, Injectable } from '@angular/core';
 import { Face } from '../model/Face';
-import { MinedOre } from '../model/MinedOre';
 import { Scan } from '../model/Scan';
 import { SelectedTile } from '../model/SelectedTile';
 
@@ -26,7 +25,6 @@ export class EventService {
    */
   public scanAdded = new EventEmitter<Scan>();
 
-  public minedOreAdded = new EventEmitter<MinedOre>();
   /**
    * emits if a longer loading process started/ended
    */
