+++ src/app/map/service/request.service.ts
@@ -6,7 +6,6 @@ import { ToastrService } from 'ngx-toastr';
 import { Scan } from '../model/Scan';
 import { Face } from '../model/Face';
 import { EventService } from './event.service';
-import { MinedOre } from '../model/MinedOre';
 
 @Injectable({
   providedIn: 'root'
@@ -73,28 +72,6 @@ export class RequestService {
     const url = `${this.defaultURL}scan`;
     this.eventService.loading.next(true);
     return this.http.post(url, scan, this.getRequestConfigObject())
-      .pipe(
-        catchError(error => {
-            this.eventService.loading.next(false);
-            console.log(error);
-            if (error.status === 0) {
-              this.toastr.error('Backend not reachable');
-            } else if (error.status === 500) {
-              this.toastr.error('Serverside error, plz report to discord channel');
-            } else if (error.status !== 400) {
-              this.toastr.error('Unexpcted error, plz report to discord channel');
-            }
-            throw error;
-          }
-        )
-      )
-      .toPromise();
-  }
-
-  saveMinedOre(minedOre: MinedOre) {
-    const url = `${this.defaultURL}minedOre`;
-    this.eventService.loading.next(true);
-    return this.http.post(url, minedOre, this.getRequestConfigObject())
       .pipe(
         catchError(error => {
             this.eventService.loading.next(false);
