+++ src/app/map/service/settings.service.ts
