--- src/app/model/User.ts
@@ -1,5 +1,0 @@
-export class User {
-    id: number;
-    username: string;
-    authdata?: string;
-}
