--- src/app/service/event.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { EventService } from './event.service';
-
-describe('EventService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: EventService = TestBed.get(EventService);
-    expect(service).toBeTruthy();
-  });
-});
