--- src/app/service/map.service.ts
@@ -1,14 +1,0 @@
-import { EventEmitter, Injectable } from '@angular/core';
-import { Face } from '../model/Face';
-import { Scan } from '../model/Scan';
-import { SelectedTile } from '../model/SelectedTile';
-
-
-@Injectable({
-  providedIn: 'root'
-})
-export class MapService {
-
-  constructor() { }
-
-}
