--- src/app/service/request.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { RequestService } from './request.service';
-
-describe('RequestService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: RequestService = TestBed.get(RequestService);
-    expect(service).toBeTruthy();
-  });
-});
